Remove paths, UMD for JST, 0.14.0
The ambiguity was killing me with the "logical" paths, an idea from Sprockets. Now require paths will be relative to the file they are in or relative to the project root. Also, extensions are mandatory so no more .js/.css ambiguity for components.

JSTs are now wrapped UMD-friendly. See the updated tests for example output.